+++ pages/dashboard.js
@@ -1,0 +1,11 @@
+import AppHeader from "../components/AppHeader";
+import Dashboard from "../components/Dashboard";
+
+export default function Dashboard() {
+  return (
+    <>
+      <AppHeader />
+      <Dashboard />
+    </>
+  );
+}
